clang-version.sh: Import new version based on cc-version.sh
clang-version.sh was rewritten upstream as cc-version.sh, which is 3x as
fast because it invokes the compiler once instead of three times.

$ hyperfine \
    --runs 200 \
    --warmup 20 \
    "./clang-version.sh clang" \
    "./cc-version.sh clang"
Benchmark 1: ./clang-version.sh clang
  Time (mean ± σ):     116.5 ms ±   1.9 ms    [User: 66.7 ms, System: 58.8 ms]
  Range (min … max):   106.5 ms … 119.5 ms    200 runs

Benchmark 2: ./cc-version.sh clang
  Time (mean ± σ):      32.8 ms ±   1.0 ms    [User: 18.2 ms, System: 15.6 ms]
  Range (min … max):    27.0 ms …  34.7 ms    200 runs

Summary
  './cc-version.sh clang' ran
    3.56 ± 0.12 times faster than './clang-version.sh clang'

This script will be used in at least one other place going forward so it
makes sense to update it to get the free performance gains (no matter
how small they are).

It has been stripped of anything not related to getting clang's version
as a five or six digit number so it looks a lot different from upstream.

Signed-off-by: Nathan Chancellor <nathan@kernel.org>